Implement draw_rectangle and better handle edge cases

+++ fb_utils.h
@@ -25,13 +25,13 @@ inline int min(int a, int b) {
 
 void *swap_buffers(Buffer *buf);
 
-void set_pixel(Buffer *buf, size_t x, size_t y, uint32_t colour);
-uint32_t get_pixel(Buffer *buf, size_t x, size_t y);
+void set_pixel(Buffer *buf, unsigned int x, unsigned int y, uint32_t colour);
+uint32_t get_pixel(Buffer *buf, unsigned int x, unsigned int y);
 
-void fill_rectangle(Buffer *buf, size_t x0, size_t y0, size_t w, size_t h, uint32_t colour);
-void draw_rectangle(Buffer *buf, size_t x0, size_t y0, size_t w, size_t h, size_t border, uint32_t colour);
+void fill_rectangle(Buffer *buf, int x0, int y0, unsigned int w, unsigned int h, uint32_t colour);
+void draw_rectangle(Buffer *buf, int x0, int y0, unsigned int w, unsigned int h, unsigned int border, uint32_t colour);
 
-void fill_arc(Buffer *buf, size_t x, size_t y, size_t r, size_t angle, uint32_t colour);
-void draw_arc(Buffer *buf, size_t x, size_t y, size_t r, size_t angle, size_t border, uint32_t colour);
+void fill_arc(Buffer *buf, int x, int y, unsigned int r, unsigned int angle, uint32_t colour);
+void draw_arc(Buffer *buf, int x, int y, unsigned int r, unsigned int angle, unsigned int border, uint32_t colour);
 
 #endif
